% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_maps_search.R
\name{get_content}
\alias{get_content}
\title{Get Outscraper content}
\usage{
get_content(parsed)
}
\arguments{
\item{parsed}{(list) the parsed JSON response from Outscraper}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get Outscraper content
}
